# dtl/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DTL API
  description: |
    API for DTL service (quote requests, appointment requests, push, settings, admin CRUD).
    See api.md for narrative documentation.
    Admin and settings endpoints are implemented as Next.js API routes; they use in-memory or
    localStorage-backed stores (no external DB yet).
  version: 1.0.0
servers:
  - url: /api
    description: Same-origin or NEXT_PUBLIC_API_URL base
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Admin login
      description: Returns JWT and sets httpOnly cookie. Use cookie or Authorization Bearer for protected routes.
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Success; Set-Cookie with JWT, body has token and expiresIn
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresIn:
                    type: integer
        "401":
          description: Invalid credentials
...
  /plati/{id}:
    get:
      tags: [Admin – Plăți]
      summary: Get plată by id
      operationId: getPlataById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plată
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientPlata"
        "404":
          description: Not found
    patch:
      tags: [Admin – Plăți]
      summary: Update plată (notes only)
      operationId: updatePlata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
      responses:
        "200":
          description: Updated plată
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientPlata"
        "404":
          description: Not found
...
  /push/notify-admin:
    post:
      tags: [Admin – Push]
      summary: Send push notification to all registered admin tokens
      description: E.g. type booking with name, date, time. Returns 503 if FCM not configured.
      operationId: pushNotifyAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifyAdminBody"
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  sent:
                    type: integer
                  failed:
                    type: integer
        "503":
          description: FCM not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /push/notify-user:
    post:
      tags: [Admin – Push]
      summary: Send push notification to user by ref
      operationId: pushNotifyUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifyUserBody"
      responses:
        "200":
          description: Sent (sent 0 or 1)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  sent:
                    type: integer
                    enum: [0, 1]
        "503":
          description: FCM not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /appointments/{id}:
    get:
      tags: [Admin – Appointments]
      summary: Get appointment by id
      operationId: adminGetAppointmentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      tags: [Admin – Appointments]
      summary: Update appointment (status and/or date/time)
      operationId: adminUpdateAppointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    delete:
      tags: [Admin – Appointments]
      summary: Delete appointment
      operationId: adminDeleteAppointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /quote-requests/{id}:
    patch:
      tags: [Admin – Quote requests]
      summary: Update quote request (e.g. status prepared)"
      operationId: adminUpdateQuoteRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: "prepared"
      responses:
        "200":
          description: Updated quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "404":
          description: Not found
  /gallery:
    get:
      tags: [Admin – Gallery]
      summary: List gallery items
      operationId: adminGetGallery
      security: []
      responses:
        "200":
          description: List of gallery items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/GalleryItem"
    post:
      tags: [Admin – Gallery]
      summary: Add gallery item
      operationId: adminAddGalleryItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemCreate"
      responses:
        "201":
          description: Created gallery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    delete:
      tags: [Admin – Gallery]
      summary: Delete gallery item
      operationId: adminDeleteGalleryItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    QuoteRequest:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        carMake:
          type: string
        carModel:
          type: string
        carYear:
          type: string
        description:
          type: string
        chassis:
          type: string
        photoUrl:
          type: string
        status:
          type: string
...
    NotifyAdminBody:
      type: object
      properties:
        type:
          type: string
          example: "booking"
        name:
          type: string
        date:
          type: string
        time:
          type: string
    NotifyUserBody:
      type: object
      required:
        - ref
      properties:
        ref:
          type: string
        title:
          type: string
          example: "Notificare"
        body:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string

    # Admin / backend entities
    Appointment:
      type: object
      description: Admin view of an appointment (lib/dummy-appointments.ts).
      properties:
        id:
          type: string
        nume:
          type: string
        telefon:
          type: string
        email:
          type: string
        data:
          type: string
          description: Date e.g. d MMM yyyy
        ora:
          type: string
          description: HH:mm
        marca:
          type: string
        model:
          type: string
        tip:
          type: string
          enum: [general, tyre, carWash]
        status:
          type: string
          enum: [Confirmat, "În așteptare"]
        descriere:
          type: string
        clientNotes:
          type: string
          description: Client's or admin-editable notes (istoric).
    AppointmentUpdate:
      type: object
      description: Partial update for appointment (status, date/time, descriere, clientNotes).
      properties:
        status:
          type: string
          enum: [Confirmat, "În așteptare"]
        data:
          type: string
          description: Date e.g. YYYY-MM-DD
        ora:
          type: string
          description: HH:mm
        descriere:
          type: string
        clientNotes:
          type: string
...
    GalleryItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        caption:
          type: string
        imageUrl:
          type: string
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
    GalleryItemCreate:
      type: object
      description: Create gallery item. imageUrl required (can be data URL from file upload); title/caption optional. id, order, createdAt server-generated.
      properties:
        imageUrl:
          type: string
          description: Required; URL or data URL (e.g. from local file upload).
        title:
          type: string
        caption:
          type: string
        order:
          type: integer
...
    ClientPlata:
      type: object
      description: Invoice/payment (lib/dummy-plati.ts).
      properties:
        id:
          type: string
        clientId:
          type: string
        nrFactura:
          type: string
        data:
          type: string
          description: YYYY-MM-DD
        suma:
          type: string
        descriere:
          type: string
        notes:
          type: string
...